requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
httplib2==0.20.2
PyJWT==2.3.0
PyMySQL==1.1.1
python-dateutil==2.8.1
python-dotenv==1.0.1
requests==2.25.1
SQLAlchemy==2.0.37
Werkzeug==3.1.3
DateTime==5.5
mysqlclient==2.1.1